refactored main activity
